Revamp Auth layout
