+++ dashboard/src/modules/components/AuthComponent/index.less
@@ -96,16 +96,4 @@
         background: transparent; /* make scrollbar transparent */
     }
 }
-.login-wrapper {
-    top: 10%;
-    .login-right-component {
-        top: 15%;
-    }
-}
-.signup-wrapper {
-    top: 5%;
 
-    .login-right-component {
-        top: 10%;
-    }
-}
