+++ dashboard/src/modules/components/TableComponent/index.jsx
@@ -146,7 +146,7 @@ const TableWithFavorite = () => {
               message="Want to see your own data?"
               link="Login or Create an account"
               onCloseMethod={onCloseLoginHint}
-              redirect={"/" + APP_ROUTES.AUTH}
+              redirect={APP_ROUTES.AUTH}
             />
           )}
           <div className="table-container">
