+++ dashboard/src/modules/containers/AuthLayout/index.less
@@ -1,0 +1,13 @@
+.login-wrapper {
+    top: 10%;
+    .login-right-component {
+        top: 15%;
+    }
+}
+.signup-wrapper {
+    top: 5%;
+
+    .login-right-component {
+        top: 10%;
+    }
+}
