--- dashboard/src/modules/containers/Dashboard/index.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Outlet } from "react-router-dom";
-
-const AuthLayout = () => {
-    return (
-        <>
-            <Outlet />
-        </>
-       
-    )
-}
-export default AuthLayout;
